streamlit-dashboard/requirements.txt
streamlit>=1.25.0
pandas>=1.3.0
google-cloud-bigquery==3.17.2
db-dtypes==1.2.0
plotly>=5.0.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
python-dotenv>=0.19.0
requests>=2.28.0